webapp/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Thomas Kientz <thomas@kientz.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
flask-restx = "^1.2.0"
faceai-bgimpact = "0.2.2"
scipy = "^1.11.4"
flask-cors = "^4.0.0"
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
